backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow==2.15.0
opencv-python==4.8.1.78
numpy==1.24.3
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2
